pyproject.toml
[tool.poetry]
name = "flask_login_webpage"
version = "0.1.0"
description = ""
authors = ["louwetienne <louwetienne@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.1"
PyJWT = "^2.4.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
